refactor : room과 관련없는 dto 이동
