+++ src/main/java/mafia/mafiatogether/game/application/GameService.java
@@ -10,10 +10,10 @@ import mafia.mafiatogether.game.domain.GameRepository;
 import mafia.mafiatogether.game.domain.Player;
 import mafia.mafiatogether.game.domain.status.EndStatus;
 import mafia.mafiatogether.game.domain.status.StatusType;
-import mafia.mafiatogether.room.application.dto.request.RoomModifyRequest;
-import mafia.mafiatogether.room.application.dto.response.RoomInfoResponse;
-import mafia.mafiatogether.room.application.dto.response.RoomResultResponse;
-import mafia.mafiatogether.room.application.dto.response.RoomStatusResponse;
+import mafia.mafiatogether.game.application.dto.request.RoomModifyRequest;
+import mafia.mafiatogether.game.application.dto.response.RoomInfoResponse;
+import mafia.mafiatogether.game.application.dto.response.RoomResultResponse;
+import mafia.mafiatogether.game.application.dto.response.RoomStatusResponse;
 import mafia.mafiatogether.room.domain.Room;
 import mafia.mafiatogether.room.domain.RoomRepository;
 import org.springframework.stereotype.Service;
