+++ src/main/java/mafia/mafiatogether/game/ui/GameController.java
@@ -4,10 +4,10 @@ import lombok.RequiredArgsConstructor;
 import mafia.mafiatogether.config.PlayerInfo;
 import mafia.mafiatogether.game.application.GameService;
 import mafia.mafiatogether.job.application.dto.PlayerInfoDto;
-import mafia.mafiatogether.room.application.dto.request.RoomModifyRequest;
-import mafia.mafiatogether.room.application.dto.response.RoomInfoResponse;
-import mafia.mafiatogether.room.application.dto.response.RoomResultResponse;
-import mafia.mafiatogether.room.application.dto.response.RoomStatusResponse;
+import mafia.mafiatogether.game.application.dto.request.RoomModifyRequest;
+import mafia.mafiatogether.game.application.dto.response.RoomInfoResponse;
+import mafia.mafiatogether.game.application.dto.response.RoomResultResponse;
+import mafia.mafiatogether.game.application.dto.response.RoomStatusResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
