+++ src/main/java/mafia/mafiatogether/job/application/PlayerService.java
@@ -12,7 +12,7 @@ import mafia.mafiatogether.job.domain.JobTarget;
 import mafia.mafiatogether.job.domain.JobTargetRepository;
 import mafia.mafiatogether.job.domain.PlayerJob;
 import mafia.mafiatogether.job.domain.PlayerJobRepository;
-import mafia.mafiatogether.room.application.dto.response.RoomNightResultResponse;
+import mafia.mafiatogether.job.application.dto.response.RoomNightResultResponse;
 import org.springframework.stereotype.Service;
 
 @Service
