+++ src/main/java/mafia/mafiatogether/job/application/dto/response/RoomNightResultResponse.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.room.application.dto.response;
+package mafia.mafiatogether.job.application.dto.response;
 
 public record RoomNightResultResponse(
     String dead
