+++ src/test/java/mafia/mafiatogether/game/ui/GameControllerTest.java
@@ -15,8 +15,8 @@ import mafia.mafiatogether.game.domain.status.StatusType;
 import mafia.mafiatogether.global.ControllerTest;
 import mafia.mafiatogether.job.application.dto.response.PlayerResponse;
 import mafia.mafiatogether.job.domain.jobtype.JobType;
-import mafia.mafiatogether.room.application.dto.response.RoomInfoResponse;
-import mafia.mafiatogether.room.application.dto.response.RoomStatusResponse;
+import mafia.mafiatogether.game.application.dto.response.RoomInfoResponse;
+import mafia.mafiatogether.game.application.dto.response.RoomStatusResponse;
 import mafia.mafiatogether.room.domain.Room;
 import mafia.mafiatogether.room.domain.RoomInfo;
 import mafia.mafiatogether.room.domain.RoomRepository;
